+++ templates/clips/app/i18n/ar-SA.ts
@@ -519,6 +519,7 @@ const messages = {
     height: "ارتفاع",
     autoplay: "التشغيل التلقائي",
     startAt: "البدء عند (ثواني)",
+    startAtTimestamp: "البدء عند {{time}}",
     embedCode: "كود التضمين",
     sharePlainTitle: "مشاركة {{title}}",
   },