+++ templates/clips/app/i18n/de-DE.ts
@@ -536,6 +536,7 @@ const messages = {
     height: "Höhe",
     autoplay: "Übersetzt: Autoplay",
     startAt: "Beginnen bei (Sekunden)",
+    startAtTimestamp: "Bei {{time}} starten",
     embedCode: "Einbettungscode",
     sharePlainTitle: "Teilen {{title}}",
   },